requirements.txt
Flask==2.3.2
torch==2.0.1
transformers==4.33.3
Pillow==10.0.0
numpy==1.24.3
lime==0.2.0.1
scikit-image==0.19.3
matplotlib==3.7.1
Werkzeug==2.3.6

